requirements.txt
aiogram==2.25.1
APScheduler==3.10.4
asyncio==3.4.3
asyncpg==0.29.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.6
aiofiles==23.1.0
